Add user entity and testing for this entity
User entity has a name, an email a password and a passwordConfirm.
In the utils.js file the dependencies for the entity isEmail,
isPasswordValid and encryptPassword are stored.
In the index.js the dependencies are injected into the user entity.
In utils.test.js the dependencies are tested and in users.test.js the
user functionality is tested.